+++ crates/zizmor/src/audit/template_injection.rs
@@ -12,7 +12,8 @@
 
 use std::sync::LazyLock;
 
-use github_actions_expressions::{BinOp, Expr, UnOp, context::ContextPattern};
+use fst::Map;
+use github_actions_expressions::Expr;
 use github_actions_models::{
     common::{Uses, expr::LoE},
     workflow::job::Strategy,
@@ -34,64 +35,28 @@ audit_meta!(
     "code injection via template expansion"
 );
 
-/// Context patterns that are believed to be always safe.
-static SAFE_CONTEXT_PATTERNS: LazyLock<Vec<ContextPattern>> = LazyLock::new(|| {
-    [
-        // The action path is always safe.
-        "github.action_path",
-        // The GitHub event name (i.e. trigger) is itself safe.
-        "github.event_name",
-        // Safe keys within the otherwise generally unsafe github.event context.
-        "github.event.after",                // hexadecimal SHA ref
-        "github.event.before",               // hexadecimal SHA ref
-        "github.event.issue.number",         // the issue's own number
-        "github.event.merge_group.base_sha", // hexadecimal SHA ref
-        "github.event.number",
-        "github.event.pull_request.base.sha", // hexadecimal SHA ref
-        "github.event.pull_request.head.sha", // hexadecimal SHA ref
-        "github.event.pull_request.head.repo.fork", // boolean
-        "github.event.pull_request.commits",  // number of commits in PR
-        "github.event.pull_request.number",   // the PR's own number
-        "github.event.workflow_run.id",
-        // Corresponds to the job ID, which is workflow-controlled
-        // but can only be [A-Za-z0-9-_].
-        // See: https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_id
-        "github.job",
-        // Information about the GitHub repository
-        "github.repository",
-        "github.repository_id",
-        "github.repositoryUrl",
-        // Information about the GitHub repository owner (account/org or ID)
-        "github.repository_owner",
-        "github.repository_owner_id",
-        // Unique numbers assigned by GitHub for workflow runs
-        "github.run_attempt",
-        "github.run_id",
-        "github.run_number",
-        // Typically something like `https://github.com`; you have bigger problems if
-        // this is attacker-controlled.
-        "github.server_url",
-        // Always a 40-char SHA-1 reference.
-        "github.sha",
-        // Like `secrets.*`: not safe to expose, but safe to interpolate.
-        "github.token",
-        // GitHub Actions-controlled local directory.
-        "github.workspace",
-        // GitHub Actions-controller runner architecture.
-        "runner.arch",
-        // Debug logging is (1) or is not (0) enabled on GitHub Actions runner.
-        "runner.debug",
-        // GitHub Actions runner operating system.
-        "runner.os",
-        // GitHub Actions temporary directory, value controlled by the runner itself.
-        "runner.temp",
-        // GitHub Actions cached tool directory, value controlled by the runner itself.
-        "runner.tool_cache",
-    ]
-    .iter()
-    .map(|s| ContextPattern::new(s).unwrap())
-    .collect()
+static CONTEXT_CAPABILITIES_FST: LazyLock<Map<&[u8]>> = LazyLock::new(|| {
+    fst::Map::new(include_bytes!(concat!(env!("OUT_DIR"), "/context-capabilities.fst")).as_slice())
+        .expect("couldn't initialize context capabilities FST")
 });
+
+enum Capability {
+    Arbitrary,
+    Structured,
+    Fixed,
+}
+
+impl Capability {
+    fn from_context(context: &str) -> Option<Self> {
+        match CONTEXT_CAPABILITIES_FST.get(context) {
+            Some(0) => Some(Capability::Arbitrary),
+            Some(1) => Some(Capability::Structured),
+            Some(2) => Some(Capability::Fixed),
+            Some(_) => unreachable!("unexpected context capability"),
+            _ => None,
+        }
+    }
+}
 
 impl TemplateInjection {
     fn script_with_location<'s>(
@@ -131,50 +96,6 @@ impl TemplateInjection {
         }
     }
 
-    /// Checks whether an expression is "safe" for the purposes of template
-    /// injection.
-    ///
-    /// In the context of template injection, a "safe" expression is one that
-    /// can only ever return a literal node (i.e. bool, number, string, etc.).
-    /// All branches/flows of the expression must uphold that invariant;
-    /// no taint tracking is currently done.
-    fn expr_is_safe(expr: &Expr) -> bool {
-        match expr {
-            Expr::Number(_) => true,
-            Expr::String(_) => true,
-            Expr::Boolean(_) => true,
-            Expr::Null => true,
-            // NOTE: Currently unreachable, since these only occur
-            // within Expr::Context and we handle that at the top-level.
-            Expr::Star | Expr::Identifier(_) | Expr::Index(_) => unreachable!(),
-            // NOTE: Some function calls may be safe, but for now
-            // we consider them all unsafe.
-            Expr::Call { .. } => false,
-            // We consider all context accesses unsafe. This isn't true,
-            // but our audit filters the safe ones later on.
-            Expr::Context { .. } => false,
-            Expr::BinOp { lhs, op, rhs } => {
-                match op {
-                    // `==` and `!=` are always safe, since they evaluate to
-                    // boolean rather than to the truthy value.
-                    BinOp::Eq | BinOp::Neq => true,
-                    // `&&` is safe if its RHS is safe, since && cannot
-                    // short-circuit.
-                    BinOp::And => Self::expr_is_safe(rhs),
-                    // We consider all other binops safe if both sides are safe,
-                    // regardless of the actual operation type. This could be
-                    // refined to check only one side with taint information.
-                    // TODO: Relax this for >/>=/</<=?
-                    _ => Self::expr_is_safe(lhs) && Self::expr_is_safe(rhs),
-                }
-            }
-            Expr::UnOp { op, .. } => match op {
-                // !expr always produces a boolean.
-                UnOp::Not => true,
-            },
-        }
-    }
-
     fn injectable_template_expressions<'s>(
         &self,
         run: &str,
@@ -187,88 +108,129 @@ impl TemplateInjection {
                 continue;
             };
 
-            if Self::expr_is_safe(&parsed) {
-                // Emit a pedantic finding for all expressions, since
-                // all template injections are code smells, even if unexploitable.
-                bad_expressions.push((
-                    expr.as_raw().into(),
-                    Severity::Unknown,
-                    Confidence::Unknown,
-                    Persona::Pedantic,
-                ));
-                continue;
-            }
+            // Emit a blanket pedantic finding for the extracted expression
+            // since any expression in a code context is a code smell,
+            // even if unexploitable.
+            bad_expressions.push((
+                expr.as_raw().into(),
+                Severity::Unknown,
+                Confidence::Unknown,
+                Persona::Pedantic,
+            ));
 
             for context in parsed.dataflow_contexts() {
-                if context.child_of("secrets") {
-                    // While not ideal, secret expansion is typically not exploitable.
-                    continue;
-                } else if SAFE_CONTEXT_PATTERNS.iter().any(|pat| pat.matches(context)) {
-                    continue;
-                } else if context.child_of("inputs") {
-                    // TODO: Currently low confidence because we don't check the
-                    // input's type. In the future, we should index back into
-                    // the workflow's triggers and exclude input expansions
-                    // from innocuous types, e.g. booleans.
-                    bad_expressions.push((
-                        context.as_str().into(),
-                        Severity::High,
-                        Confidence::Low,
-                        Persona::default(),
-                    ));
-                } else if let Some(env) = context.pop_if("env") {
-                    let env_is_static = step.env_is_static(env);
+                // Try and turn our context into a pattern for
+                // matching against the FST.
+                match context.as_pattern().as_deref() {
+                    Some(context_pattern) => {
+                        // Try and get the pattern's capability from our FST.
+                        match Capability::from_context(context_pattern) {
+                            // Fixed means no meaningful injectable structure.
+                            Some(Capability::Fixed) => continue,
+                            // Structured means some attacker-controllable
+                            // structure, but not fully arbitrary.
+                            Some(Capability::Structured) => {
+                                bad_expressions.push((
+                                    context.as_str().into(),
+                                    Severity::Medium,
+                                    Confidence::High,
+                                    Persona::default(),
+                                ));
+                            }
+                            // Arbitrary means the context's expansion is
+                            // fully attacker-controllable.
+                            Some(Capability::Arbitrary) => {
+                                bad_expressions.push((
+                                    context.as_str().into(),
+                                    Severity::High,
+                                    Confidence::High,
+                                    Persona::default(),
+                                ));
+                            }
+                            None => {
+                                // Without a FST match, we fall back on heuristics.
+                                if context.child_of("secrets") {
+                                    // While not ideal, secret expansion is typically not exploitable.
+                                    continue;
+                                } else if context.child_of("inputs") {
+                                    // TODO: Currently low confidence because we don't check the
+                                    // input's type. In the future, we should index back into
+                                    // the workflow's triggers and exclude input expansions
+                                    // from innocuous types, e.g. booleans.
+                                    bad_expressions.push((
+                                        context.as_str().into(),
+                                        Severity::High,
+                                        Confidence::Low,
+                                        Persona::default(),
+                                    ));
+                                } else if let Some(env) = context.pop_if("env") {
+                                    let env_is_static = step.env_is_static(env);
 
-                    if !env_is_static {
+                                    if !env_is_static {
+                                        bad_expressions.push((
+                                            context.as_str().into(),
+                                            Severity::Low,
+                                            Confidence::High,
+                                            Persona::default(),
+                                        ));
+                                    }
+                                } else if context.child_of("github") {
+                                    // TODO: Filter these more finely; not everything in the event
+                                    // context is actually attacker-controllable.
+                                    bad_expressions.push((
+                                        context.as_str().into(),
+                                        Severity::High,
+                                        Confidence::High,
+                                        Persona::default(),
+                                    ));
+                                } else if context.child_of("matrix") || context == "matrix" {
+                                    if let Some(Strategy { matrix, .. }) = step.strategy() {
+                                        let matrix_is_static = match matrix {
+                                            // The matrix is generated by an expression, meaning
+                                            // that it's trivially not static.
+                                            Some(LoE::Expr(_)) => false,
+                                            // The matrix may expand to static values according to the context
+                                            Some(inner) => models::Matrix::new(inner)
+                                                .expands_to_static_values(context.as_str()),
+                                            // Context specifies a matrix, but there is no matrix defined.
+                                            // This is an invalid workflow so there's no point in flagging it.
+                                            None => continue,
+                                        };
+
+                                        if !matrix_is_static {
+                                            bad_expressions.push((
+                                                context.as_str().into(),
+                                                Severity::Medium,
+                                                Confidence::Medium,
+                                                Persona::default(),
+                                            ));
+                                        }
+                                    }
+                                    continue;
+                                } else {
+                                    // All other contexts are typically not attacker controllable,
+                                    // but may be in obscure cases.
+                                    bad_expressions.push((
+                                        context.as_str().into(),
+                                        Severity::Informational,
+                                        Confidence::Low,
+                                        Persona::default(),
+                                    ));
+                                }
+                            }
+                        }
+                    }
+                    None => {
+                        // If we couldn't turn the context into a pattern,
+                        // we almost certainly have something like
+                        // `call(...).foo.bar`.
                         bad_expressions.push((
                             context.as_str().into(),
-                            Severity::Low,
-                            Confidence::High,
+                            Severity::Informational,
+                            Confidence::Low,
                             Persona::default(),
                         ));
                     }
-                } else if context.child_of("github") {
-                    // TODO: Filter these more finely; not everything in the event
-                    // context is actually attacker-controllable.
-                    bad_expressions.push((
-                        context.as_str().into(),
-                        Severity::High,
-                        Confidence::High,
-                        Persona::default(),
-                    ));
-                } else if context.child_of("matrix") || context == "matrix" {
-                    if let Some(Strategy { matrix, .. }) = step.strategy() {
-                        let matrix_is_static = match matrix {
-                            // The matrix is generated by an expression, meaning
-                            // that it's trivially not static.
-                            Some(LoE::Expr(_)) => false,
-                            // The matrix may expand to static values according to the context
-                            Some(inner) => models::Matrix::new(inner)
-                                .expands_to_static_values(context.as_str()),
-                            // Context specifies a matrix, but there is no matrix defined.
-                            // This is an invalid workflow so there's no point in flagging it.
-                            None => continue,
-                        };
-
-                        if !matrix_is_static {
-                            bad_expressions.push((
-                                context.as_str().into(),
-                                Severity::Medium,
-                                Confidence::Medium,
-                                Persona::default(),
-                            ));
-                        }
-                    }
-                    continue;
-                } else {
-                    // All other contexts are typically not attacker controllable,
-                    // but may be in obscure cases.
-                    bad_expressions.push((
-                        context.as_str().into(),
-                        Severity::Informational,
-                        Confidence::Low,
-                        Persona::default(),
-                    ));
                 }
             }
         }
@@ -331,56 +293,27 @@ impl Audit for TemplateInjection {
 
 #[cfg(test)]
 mod tests {
-    use super::{Expr, TemplateInjection};
+    use crate::audit::template_injection::Capability;
 
     #[test]
-    fn test_expr_is_safe() {
-        let cases = &[
-            // Literals are always safe.
-            ("true", true),
-            ("false", true),
-            ("1.0", true),
-            ("null", true),
-            ("'some string'", true),
-            // negation is always safe.
-            ("!true", true),
-            ("!some.context", true),
-            // == / != are always safe, even if their hands are not.
-            ("true == true", true),
-            ("'true' == true", true),
-            ("some.context == true", true),
-            ("contains(some.context, 'foo') != true", true),
-            // || is safe if both hands are safe.
-            ("true || true", true),
-            ("some.context || true", false),
-            ("true || some.context", false),
-            // && is true if the RHS is safe.
-            ("true && true", true),
-            ("some.context && true", true),
-            ("true && other.context", false),
-            ("some.context && other.context", false),
-            // Index ops and function calls are unsafe.
-            ("some.context[0]", false),
-            ("some.context[*]", false),
-            ("someFunction()", false),
-            ("fromJSON(some.context)", false),
-            ("toJSON(fromJSON(some.context))", false),
-            // Context accesses are unsafe.
-            ("some.context", false),
-            ("some.context.*.something", false),
-            // More complicated cases:
-            ("some.condition && '--some-arg' || ''", true),
-            ("some.condition && some.context || ''", false),
-            ("some.condition && '--some-arg' || some.context", false),
-            (
-                "(github.actor != 'github-actions[bot]' && github.actor) || 'BrewTestBot'",
-                false,
+    fn test_capability_from_context() {
+        assert!(matches!(
+            Capability::from_context("github.event.workflow_run.triggering_actor.login"),
+            Some(Capability::Arbitrary)
+        ));
+        assert!(matches!(
+            Capability::from_context(
+                "github.event.workflow_run.triggering_actor.organizations_url"
             ),
-        ];
-
-        for (case, safe) in cases {
-            let expr = Expr::parse(case).unwrap();
-            assert_eq!(TemplateInjection::expr_is_safe(&expr), *safe, "{expr:#?}");
-        }
+            Some(Capability::Structured)
+        ));
+        assert!(matches!(
+            Capability::from_context("github.event.type.is_enabled"),
+            Some(Capability::Fixed)
+        ));
+        assert!(matches!(
+            Capability::from_context("runner.arch"),
+            Some(Capability::Fixed)
+        ));
     }
 }