+++ docs/release-notes.md
@@ -13,6 +13,11 @@ of `zizmor`.
 
 * `zizmor` now supports generating completions for Nushell (#838)
 
+### Bug Fixes 🐛
+
+* The [insecure-commands] now correctly detects different truthy
+  values in `ACTIONS_ALLOW_UNSECURE_COMMANDS` (#840)
+
 ## v1.8.0
 
 ### Announcements 📣
@@ -805,3 +810,4 @@ This is one of `zizmor`'s bigger recent releases! Key enhancements include:
 [stale-action-refs]: ./audits.md#stale-action-refs
 [unsound-contains]: ./audits.md#unsound-contains
 [unpinned-images]: ./audits.md#unpinned-images
+[insecure-commands]: ./audits.md#insecure-commands