+++ docs/release-notes.md
@@ -13,6 +13,15 @@ of `zizmor`.
 
 * `zizmor` now supports generating completions for Nushell (#838)
 
+### Enhancements 🌱
+
+* The [template-injection] audit has been rewritten, and is now significantly
+  more precise and general over contexts supplied via GitHub's webhook
+  payloads (i.e. `github.event.*`) (#745)
+* The [template-injection] audit now correctly emits pedantic findings
+  in a blanket manner, rather than filtering them based on the presence
+  of other findings (#745)
+
 ### Bug Fixes 🐛
 
 * The [insecure-commands] now correctly detects different truthy