+++ docs/release-notes.md
@@ -18,14 +18,14 @@ of `zizmor`.
 * The [template-injection] audit has been rewritten, and is now significantly
   more precise and general over contexts supplied via GitHub's webhook
   payloads (i.e. `github.event.*`) (#745)
-* The [template-injection] audit now correctly emits pedantic findings
-  in a blanket manner, rather than filtering them based on the presence
-  of other findings (#745)
 
 ### Bug Fixes 🐛
 
 * The [insecure-commands] now correctly detects different truthy
   values in `ACTIONS_ALLOW_UNSECURE_COMMANDS` (#840)
+* The [template-injection] audit now correctly emits pedantic findings
+  in a blanket manner, rather than filtering them based on the presence
+  of other findings (#745)
 
 ## v1.8.0