+++ docs/release-notes.md
@@ -29,6 +29,8 @@ of `zizmor`.
 * The [template-injection] audit now correctly emits pedantic findings
   in a blanket manner, rather than filtering them based on the presence
   of other findings (#745)
+* CLI: Fixed a misleading error message when `zizmor` is used with
+  a GitHub host other than `github.com` (#863)
 
 ## v1.8.0